+++ src/offboard_control_interfaces/msg/Drone3Info.msg
