+++ src/offboard_control_interfaces/package.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<package format="3">
+  <name>offboard_control_interfaces</name>
+  <version>0.0.1</version>
+  <description>Custom message interfaces for offboard drone control</description>
+  <maintainer email="l.jordan@student.tudelft.nl">Lauren Jordan</maintainer>
+  <license>MIT</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <build_depend>rosidl_default_generators</build_depend>
+  <exec_depend>rosidl_default_runtime</exec_depend>
+
+  <depend>builtin_interfaces</depend>
+  <depend>std_msgs</depend>
+
+  <member_of_group>rosidl_interface_packages</member_of_group>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
