+++ src/offboard_control_pkg/package.xml
@@ -7,25 +7,24 @@
   <maintainer email="l.jordan@student.tudelft.nl">Lauren Jordan</maintainer>
   <license>MIT</license>
 
-  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_python</buildtool_depend>
 
   <build_depend>rclpy</build_depend>
   <build_depend>px4_msgs</build_depend>
   <build_depend>numpy</build_depend>
-  <build_depend>rosidl_default_generators</build_depend>
-  <build_depend>std_msgs</build_depend>
- 
+
   <exec_depend>rclpy</exec_depend>
   <exec_depend>px4_msgs</exec_depend>
   <exec_depend>numpy</exec_depend>
-  <exec_depend>rosidl_default_runtime</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
 
   <test_depend>ament_copyright</test_depend>
   <test_depend>ament_flake8</test_depend>
   <test_depend>ament_pep257</test_depend>
   <test_depend>python3-pytest</test_depend>
 
-  <member_of_group>rosidl_interface_packages</member_of_group>
+  <depend>offboard_control_interfaces</depend>
 
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
 </package>
